% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PPV.R
\name{calculate_classic_ppv}
\alias{calculate_classic_ppv}
\title{Calculate PPV using simple algorithm}
\usage{
calculate_classic_ppv(
  beats,
  PP_col = "PP",
  time_col = 1,
  resp_rate = NULL,
  resp_len = NULL,
  average_over_n = 3,
  average_fun = mean,
  plot_diagnostic = FALSE
)
}
\arguments{
\item{beats}{Data frame of beats incl. pulse pressure}

\item{PP_col}{Column containing pulse pressure}

\item{time_col}{Column containing time (datetime or seconds)}

\item{resp_rate}{Either resp_rate (breaths/min) or resp_len (seconds)}

\item{resp_len}{Either resp_rate (breaths/min) or resp_len (seconds)}

\item{average_over_n}{Number of max-min pairs (resp cycles) to average over}

\item{average_fun}{Function used for averaging}

\item{plot_diagnostic}{Show diagnostic plot}
}
\value{

}
\description{
This function attempt to calculate PPV as described by De Backer, 2009.
}
