% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_processing.R
\name{find_peaks}
\alias{find_peaks}
\title{Peak detection algorithm from https://github.com/stas-g/findPeaks}
\usage{
find_peaks(x, m = 3, na.ignore = FALSE)
}
\arguments{
\item{x}{signal vector}

\item{m}{Number of points of both sides of the peak with a lower value.
2m = width of peak}

\item{na.ignore}{Ignore NA values in peak calculation}
}
\description{
To find valleys, use -x
}
