% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PPV.R
\name{find_PP_minmax}
\alias{find_PP_minmax}
\title{Find respiratory min and max in pulse pressure time series}
\usage{
find_PP_minmax(
  beats,
  PP_col = "PP",
  time_col = 1,
  resp_rate = NULL,
  resp_len = NULL
)
}
\arguments{
\item{beats}{Data frame of beats incl. pulse pressure}

\item{PP_col}{Column containing pulse pressure}

\item{time_col}{Column containing time (datetime or seconds)}

\item{resp_rate, resp_len}{Either resp_rate (breaths/min) or resp_len (seconds)}
}
\value{

}
\description{
Find respiratory min and max in pulse pressure time series
}
