% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_waveforms.R
\name{gen_annotation_index}
\alias{gen_annotation_index}
\title{Add indexing of signal time in relation to some recurring event.}
\usage{
gen_annotation_index(...)
}
\description{
Index is the time since the most recent occurrence of some event e.g. QRS complex
}
