% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_tools.R
\name{PP_gam_analysis}
\alias{PP_gam_analysis}
\title{Analyse Pulse Pressure data using a GAM
Respiratory cycle length is found by using the value that minimizes model
REML score}
\usage{
PP_gam_analysis(
  data,
  plot = FALSE,
  return_GAM = FALSE,
  PP_col = "PP",
  time_col = "time",
  search_interval = c(1.5, 7)
)
}
\arguments{
\item{data}{Data frame of timings and pulse pressures of beats}

\item{plot}{Diagnostic GAM plot usign gratia::draw}

\item{return_GAM}{Return fitted GAM instead of data frame}

\item{PP_col}{Column with pulse pressure values}

\item{time_col}{Columns with times}

\item{search_interval}{Respiratory cycle length intervals to search}
}
\value{
Data frame of respiratory cycle length, respiratory rate and PPV.
}
\description{
Analyse Pulse Pressure data using a GAM
Respiratory cycle length is found by using the value that minimizes model
REML score
}
