% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{record_time_to_hms}
\alias{record_time_to_hms}
\title{Convert record datetime to HMS}
\usage{
record_time_to_hms(rec, as_datetime = FALSE)
}
\arguments{
\item{rec}{(nested) list of signals}
}
\description{
Convert all timestamps to HMS (seconds since midnight)
}
